refactor: 코드 전반의 리팩토링
