+++ client/src/components/TaskModal/ImportanceInput.tsx
@@ -7,11 +7,11 @@ const ImportanceInput = () => {
   return (
     <S.Rating>
       <div>
-        <S.Star isClicked={importance >= 1} onClick={() => setImportance(1)} />
-        <S.Star isClicked={importance >= 2} onClick={() => setImportance(2)} />
-        <S.Star isClicked={importance >= 3} onClick={() => setImportance(3)} />
-        <S.Star isClicked={importance >= 4} onClick={() => setImportance(4)} />
-        <S.Star isClicked={importance >= 5} onClick={() => setImportance(5)} />
+        <S.Star clicked={importance >= 1} onClick={() => setImportance(1)} />
+        <S.Star clicked={importance >= 2} onClick={() => setImportance(2)} />
+        <S.Star clicked={importance >= 3} onClick={() => setImportance(3)} />
+        <S.Star clicked={importance >= 4} onClick={() => setImportance(4)} />
+        <S.Star clicked={importance >= 5} onClick={() => setImportance(5)} />
       </div>
     </S.Rating>
   );
