+++ client/src/index.css
@@ -20,4 +20,5 @@ code {
   --color-gray4: lightgray;
   --color-gray5: #a3a3a3;
   --color-gray6: gray;
+  --font-title: 'Press Start 2P', cursive;
 }
